fixed everything maybe
