+++ public/index.html
@@ -20,7 +20,8 @@
   <link rel="stylesheet" type="text/css" media="screen" href="./modal.css" />
 
   <!-- FontAwesome CSS -->
-  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.0.13/css/all.css" integrity="sha384-DNOHZ68U8hZfKXOrtjWvjxusGo9WQnrNx2sqG0tfsghAvtVlRW3tvkXWZh58N9jp" crossorigin="anonymous">
+  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.0.13/css/all.css" integrity="sha384-DNOHZ68U8hZfKXOrtjWvjxusGo9WQnrNx2sqG0tfsghAvtVlRW3tvkXWZh58N9jp"
+    crossorigin="anonymous">
 
 
 
@@ -43,13 +44,17 @@
 
     <div class="buttons">
       <div class="add-card button" type="button">
-        <h2>+ <i class="fas fa-chalkboard"></i></h2>
+        <h2>+
+          <i class="fas fa-chalkboard"></i>
+        </h2>
       </div>
-  
+
       <div class="add-tag button">
-        <h2>+ <i class="fas fa-tag"></i></h2>
+        <h2>+
+          <i class="fas fa-tag"></i>
+        </h2>
       </div>
-  
+
       <div class="login button">
         <h2>
           <a href="http://localhost:8080/auth/github">Login</a>
@@ -78,7 +83,7 @@
             const element = array[index];
         }
       </textarea>
-    
+
       <div class="buttons">
         <button class="js-copy-button">Copy</button>
         <button class="js-edit-button">Edit</button>
@@ -86,12 +91,12 @@
       </div>
 
       <div class="votes">
-          <i class="fas fa-thumbs-up"></i>
-          <span class="js-thumbs-up">3</span>
-          <i class="fas fa-thumbs-down"></i>
-          <span class="js-thumbs-down">1</span>
+        <i class="fas fa-thumbs-up"></i>
+        <span class="js-thumbs-up">3</span>
+        <i class="fas fa-thumbs-down"></i>
+        <span class="js-thumbs-down">1</span>
       </div>
-      
+
     </div>
 
     <div class="card">
@@ -107,21 +112,21 @@ for (let index = 0; index < array.length; index++) {
     const element = array[index];
 }
     </textarea>
-  
-    <div class="buttons">
-      <button class="js-copy-button">Copy</button>
-      <button class="js-edit-button">Edit</button>
-      <button class="js-delete-button">Delete</button>
-    </div>  
 
-    <div class="votes">
+      <div class="buttons">
+        <button class="js-copy-button">Copy</button>
+        <button class="js-edit-button">Edit</button>
+        <button class="js-delete-button">Delete</button>
+      </div>
+
+      <div class="votes">
         <i class="fas fa-thumbs-up"></i>
         <span class="js-thumbs-up">3</span>
         <i class="fas fa-thumbs-down"></i>
         <span class="js-thumbs-down">1</span>
+      </div>
+
     </div>
-    
-  </div>
 
     <div class="card">
       <div class="top">
@@ -136,21 +141,21 @@ for (let index = 0; index < array.length; index++) {
     const element = array[index];
 }
     </textarea>
-  
-    <div class="buttons">
-      <button class="js-copy-button">Copy</button>
-      <button class="js-edit-button">Edit</button>
-      <button class="js-delete-button">Delete</button>
-    </div>  
 
-    <div class="votes">
+      <div class="buttons">
+        <button class="js-copy-button">Copy</button>
+        <button class="js-edit-button">Edit</button>
+        <button class="js-delete-button">Delete</button>
+      </div>
+
+      <div class="votes">
         <i class="fas fa-thumbs-up"></i>
         <span class="js-thumbs-up">3</span>
         <i class="fas fa-thumbs-down"></i>
         <span class="js-thumbs-down">1</span>
+      </div>
+
     </div>
-    
-  </div>
 
     <div class="card">
       <div class="top">
@@ -165,62 +170,65 @@ for (let index = 0; index < array.length; index++) {
     const element = array[index];
 }
     </textarea>
-  
-    <div class="buttons">
-      <button class="js-copy-button">Copy</button>
-      <button class="js-edit-button">Edit</button>
-      <button class="js-delete-button">Delete</button>
-    </div>  
 
-    <div class="votes">
+      <div class="buttons">
+        <button class="js-copy-button">Copy</button>
+        <button class="js-edit-button">Edit</button>
+        <button class="js-delete-button">Delete</button>
+      </div>
+
+      <div class="votes">
         <i class="fas fa-thumbs-up"></i>
         <span class="js-thumbs-up">3</span>
         <i class="fas fa-thumbs-down"></i>
         <span class="js-thumbs-down">1</span>
+      </div>
+
     </div>
-    
-  </div>
 
-    <!-- BOILERPLATE MODAL -->
-  <div class="modal" id="bpModal">
-  <div class="modal-content">
-    <span class ="close">&times;</span>
-    <h1>Submit your own boilerplate!</h1>
-
-    <label for="title">Title: </label>
-    <input type="text" id="title" name="title">
-    <label for="description">Description: </label>
-    <input type="text" id="description" name="description">
-    <label for="lang">Language (e.g. "Javascript"): </label>
-    <input type="text" id="lang" name="lang">
-    <label for="content">Your code: </label>
-    <input type="text" id="content" name="content">
-    <label for="tag" class="tag-label">Tag: </label>
-    <select id="tag-select">
-      <option value="tag1">Tags go here</option>
-      <option value="tag2">Tags go here</option>    </select>
-
-    <button class="post-boilerplate">Submit</button>
-  </div>
-</div>
-
-<div class="modalTag" id="tagModal">
-    <div class="modal-content">
-      <span class="closeTag">&times;</span>
-      <h1>Submit your own tags!</h1>
-  
-      <label for="tag-title">Title: </label>
-      <input type="text" id="tag-title" name="tag-title">
-      <label for="color">Color: </label>
-      <input type="text" id="color" name="color" class="jscolor">
-  
-      <button class="post-tag">Submit</button>
-    </div>
-  </div>
 
 
 
   </main>
+
+  <!-- BOILERPLATE MODAL -->
+  <div class="modal" id="bpModal">
+    <div class="modal-content">
+      <span class="close">&times;</span>
+      <h1>Submit your own boilerplate!</h1>
+
+      <label for="title">Title: </label>
+      <input type="text" id="title" name="title">
+      <label for="description">Description: </label>
+      <input type="text" id="description" name="description">
+      <label for="lang">Language (e.g. "Javascript"): </label>
+      <input type="text" id="lang" name="lang">
+      <label for="content">Your code: </label>
+      <input type="text" id="content" name="content">
+      <label for="tag" class="tag-label">Tag: </label>
+      <select id="tag-select">
+        <option value="tag1">Tags go here</option>
+        <option value="tag2">Tags go here</option>
+      </select>
+
+      <button class="post-boilerplate">Submit</button>
+    </div>
+  </div>
+
+  <div class="modalTag" id="tagModal">
+    <div class="modal-content">
+      <span class="closeTag">&times;</span>
+      <h1>Submit your own tags!</h1>
+
+      <label for="tag-title">Title: </label>
+      <input type="text" id="tag-title" name="tag-title">
+      <label for="color">Color: </label>
+      <input type="text" id="color" name="color" class="jscolor">
+
+      <button class="post-tag">Submit</button>
+    </div>
+  </div>
+
 
 
   <footer>
@@ -232,13 +240,14 @@ for (let index = 0; index < array.length; index++) {
     <!-- Underscore.JS -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/underscore.js/1.9.0/underscore-min.js"></script>
 
-    <!-- Widgets -->
-    <script src="./widgets.js"></script>
+
 
     <!-- AJAX -->
     <script src="./ajax.js"></script>
 
-    <script src="./contents.js"></script>
+    <!-- Widgets -->
+    <script src="./widgets.js"></script>
+
 
     <!-- <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
   <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script> -->
