+++ routes/api-routes.js
@@ -6,6 +6,8 @@
 
 // Requiring our models
 var db = require("../models");
+var Sequelize = require("sequelize");
+const Op = Sequelize.Op;
 
 // Routes =============================================================
 const routing = {
@@ -24,6 +26,27 @@ const routing = {
       }
     }).then(bpOne => {
       return bpOne;
+    });
+  },
+
+  // search function
+  searchBoilerplates: searchQuery => {
+    return db.Boilerplate.findAll({
+      where: {
+        [Op.or]: [
+          {
+            title: {
+              [Op.like]: "%" + searchQuery + "%"
+            }
+          },
+          {
+            description: {
+              [Op.like]: "%" + searchQuery + "%"
+            }
+          }
+        ]
+      },
+      include: [{ model: db.User }, { model: db.Tag }]
     });
   },
   // due to its dependence to db connection, this function is tested in api.int.test.js
@@ -147,6 +170,13 @@ const routing = {
     //       res.json(bp);
     //     });
     // });
+
+    app.post("/api/search", (req, res) => {
+      console.log("The query is " + req.body.searchQuery);
+      routing
+        .searchBoilerplates(req.body.searchQuery)
+        .then(result => res.json(result));
+    });
 
     app.post("/api/boilerplates", function(req, res) {
       // create takes an argument of an object describing the item we want to insert
